src/Evaluation.cpp: Expanded subpair subterm binding for sequence prefix binding.
The manually inlining is not only (hopefully) more efficient, but also cleanup the path of the further optimization of sigil reusing. Since this is the only path that not referencing the temporary object, other paths can reduce one more argument to specify the option by the sigil instead of 'ref_temp', because in all other paths these two are equivalent.

Signed-off-by: He Bin <hebin@uniontech.com>